requirements.txt
# Sensor Anomaly Explanation Generator — Python Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
anthropic==0.39.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-multipart==0.0.12
